```             inl {out bck} = final label input s             state ```
As expected the cost function itself is negligible. Well, 0.05s is not a little, but I'll ignore it. This is just Spiral being slow at compilation.

```
Struct.foldr (inl {bck} _ -> Struct.foldr (inl bck _ -> bck()) bck ()) network ()
```

Oh, even with this commented in it is still only 1.73s. What takes it to 2.2s would be optimization. Damn man.

But I did cut it quite a bit now...5 layers use to take 3s previously.

Before I move on to optimizing the kfac function, what I will do is take a look if there is anything I can trim in the backward passes.