``` inl alpha = Math.pow (one - rate) k ```
It is probably because of this. Let me switch to `**`.

Yeah, that was it.

```
            inl l = Struct.map' (indiv >> dyn) l
            join
            Struct.iter bck l
            Struct.iter (inl {streams=x} -> Tuple.iter s.data.stream.wait_on x) l
```

This does speed things up from 1.7 to 1.65, but I'd rather have the checks be done at compile time. Forget this.

This improvement might just be noise.

Let me bring back the testing pass. I suppose I should look into `ln_relu` after that.