Adding each layer to the char-RNN test adds 0.5s to the total time.
Here I will be able to find some ways of speeding it up, definitely. But I am not sure I feel like messing with this right now. I had enough for one day.

Is there even a point in speeding up the second test? How about I just go to the poker game instead?

No, I need to check it out. I need to do this so I can start the next phase with a clear head. I'll burn tomorrow for the sake of getting this done. What I set out to do, I do eventually.

That having said, I am having a new feeling right now. I am starting to loathe the entire field of RL for not giving me what I want. Is reliability on these shitty toy games so much to ask? These are the systems I am looking to entrust my future to. Is this the most they can do? Have the geniuses working to spearhead it done nothing but jack off for the last few decades?

It is really not looking good. I really will have to go Bayesian if I want to get anywhere with ML in the end at this rate.