"2:20pm. No.
I am done with speeding up `learning2`. I manged to bring it down to 3.7 from 5.5 for this net:

```
        rnn =
            rnn 128,
            rnn 128,
            rnn 128,
            rnn 128,
            rnn 128,
            rnn 128,
            linear size.hot
```

That still leaves the question why for every layer added the compilation goes up by 0.2s, but I do not want to mess with this anymore.

2:25pm. Right now I have one more thing left to figure out...

Wait actually, why do I have to put it in the outer layer? Couldn't I extract the weights first and iterate over them? Or maybe...

```
    inl standard s x =
        Struct.iter (function
            | {optimize weights} -> join
                inl weights = Struct.map' (inl x -> x.data) weights
                optimize {weights} s
            | {weights} -> join
                Struct.iter (function {weight} -> sgd s weight.data | _ -> ()) weights
            ) x
```

Let me do it like this as it is more roboust this way.

Now before I declare work done, I need to figure out why the runtime performance in learning2 because so diffent ever since I finished the redesign."