Whops, I did it in the RNN by accident. Let me do it properly now.
```
The learning rate is 2 ** -13
Starting timing for: iteration 0
The time was 00:00:06.5900052 for: iteration 0
Training: 2.41249017007923
The learning rate is 2 ** -13
Starting timing for: iteration 1
The time was 00:00:06.5429372 for: iteration 1
Training: 1.89754001980456
The learning rate is 2 ** -13
Starting timing for: iteration 2
The time was 00:00:06.5640577 for: iteration 2
Training: 1.79902508215267
The learning rate is 2 ** -13
Starting timing for: iteration 3
The time was 00:00:06.5870332 for: iteration 3
Training: 1.74826675707557
The learning rate is 2 ** -13
Starting timing for: iteration 4
The time was 00:00:06.5629752 for: iteration 4
Training: 1.71575371725896
```

Yeah, I think this is similar to the performance I had previously, but 6.5s per iteration does feel kind of high. It should be more like 5.2s. I am going to have to go to the earlier commits in order to check this out.