"2:40pm. I am not sure which commit to pick. Let me try the 23rd of last month.
2:45pm. I am suprised that the thing from last month compiles in 2.6s. I thought it would be inefficient. What is surprising that it does that well without the latest optimizations.

```
The learning rate is 2 ** -13
Starting timing for: iteration 0
The time was 00:00:06.2805493 for: iteration 0
Training: 2.31374355047421
The learning rate is 2 ** -13
Starting timing for: iteration 1
The time was 00:00:06.2560599 for: iteration 1
Training: 1.85511121753713
The learning rate is 2 ** -13
Starting timing for: iteration 2
The time was 00:00:06.2406872 for: iteration 2
Training: 1.76020719039747
The learning rate is 2 ** -13
Starting timing for: iteration 3
The time was 00:00:06.2602744 for: iteration 3
Training: 1.71031681014745
The learning rate is 2 ** -13
Starting timing for: iteration 4
The time was 00:00:06.0664425 for: iteration 4
Training: 1.67788337918875
```

Here is the permance I am getting with it. I think I understand why this is happening.

Two things could account for the fact that the new version is slower.

```
apply = inl {weights input} -> matmultb (input, weights.input) weights.bias >>= activation
```

Namely, that I am using the old matmultb which does no updating of covariance matrices.

And that I am not concatting the biases to the input. That is no doubt an overhead.

I guess my memory must be wrong that it was taking 5.2s. Maybe that was before KFAC. Let me look through the journal.

Agh...I can't find anything from `learning2`. Let me go back through the commits instead. I'll check from a month ago and then stop looking into it so much.

Actually, I do think that KFAC added an extra second to that score. I recall it now doing that. Yeah, that makes sense."