I am almost ready to start. Today I will dedicate the time to getting the poker game back online. Yesterday I only meant to read the Bayesian RL book for a bit, but I managed to go through a quarter of it.
It is excellent. I learned more in an hour of it than during the entire month of the PGM course back in 2015. I can do this.

If I can master probabilistic programming and merge it with deep learning, I can meet all of my goals. I said I would drop Spiral for a bit, but I wonder...if I was the kind of guy to relly on other's work, I'd have done it long ago. But I won't make it a principle that I have to use Spiral for this next part.

What I want to is deal with this now so I can leave the library in a pristine condition as I move to the next step. Let me finish the chilling and then I will start work.