``` Error trace on line: 261, column: 15 in file "CudaAD".             | x ->               ^ Error trace on line: 262, column: 17 in file "CudaAD".                 Struct.map (function                 ^ Error trace on line: 266, column: 24 in file "CudaAD".                     ) (View.unzip x)                        ^ Pattern miss error. The argument is {bodies=<function>; dim=[]} ```
Hmmm...

```
Error trace on line: 143, column: 39 in file "Learning".
                Struct.map' (inl init cur ->
                                      ^
Error trace on line: 144, column: 21 in file "Learning".
                    (init cur >>= succ) .out
                    ^
Error trace on line: 145, column: 21 in file "Learning".
                    |> Struct.map (function
                    ^
Error trace on line: 144, column: 22 in file "Learning".
                    (init cur >>= succ) .out
                     ^
```

I am going to have to investigate this. I am not surprised that I did not get everything in CudaAD.

Hmmm, this is in segmented init.

Is there something wrong with `concat` again?