``` Error trace on line: 415, column: 29 in file "Learning".     inl softmax_body temp k input =                             ^ Error trace on line: 416, column: 17 in file "Learning".         inl x = k.block.map (inl x -> x / temp) (k.block.load input)                 ^ Error trace on line: 528, column: 22 in file "CudaKernel".             load=inl (!zip tns) -> create_items (inl {i} -> tns i .get)                      ^ Cannot find a member named bodies inside the module. ```
Damn, this is taking a while. Debugging can never go according to schedule. It is good that I stopped when I decided to yesterday.

Now I am wondering whether the old tests still work. Let me check it out and I will take a break. I had enough of this for the time being.