Focus. Stop surfing 4chan. Deal with the error. You can do this.
```
{dim={from=lit 0i64; near_to=lit 1i64}; from=var (int64)}
var (int64)
{dim={from=lit 0i64; near_to=lit 1i64}; from=var (int64)}
var (int64)
{dim={from=lit 0i64; near_to=lit 1i64}; from=var (int64)}
[lit 0i64, lit 0i64]
{dim={from=lit 0i64; near_to=lit 1i64}; from=var (int64)}
[lit 0i64, lit 0i64]
{dim={from=lit 0i64; near_to=lit 1i64}; from=var (int64)}
[lit 0i64, lit 0i64]
```

Ok, this is unexpected. What the hell is going on here?

```
{dim={from=lit 0i64; near_to=lit 1i64}; from=var (int64)}
[lit 0i64, lit 0i64]
{cur=[lit 1i64, lit 1i64]; to={R=var (array_cuda_local (float32)); error=[]}}
[lit 1i64, lit 1i64]
```

Here is it a mit cleaned up. This error is massively convusing me. In the tuple conversion how can it possibly be converting to two zeroes when the dim is 1d?

Let me first start by figuring that out.