3:50pm. Ah, I am overdue on chores. I never expected this would take so long.
I seem to be going in the wrong direction. I am introducing complexity rather than resolving it.

4:15pm. Done with chores.

It is time to resume. I admit, this is going a lot worse than I'd imagined it would be. I honestly, though there would not be much work to get the poker agent to run, but now it feels like I won't even finish today.

...Ok. I see now how this error could occur in `segmented_init`. It is an easy fix.

```
inl link_adjoint cur {from to} =
    Struct.iter2 (inl from to ->
        match to with
        | () -> ()
        | _ -> to 0 <- from .view cur .get
        ) from to
    {
    out=()
    bck=inl _ ->
        Struct.iter2 (inl from to ->
            match to with
            | () -> ()
            | _ -> from .view cur .set (to 0)
            ) from to
    }
```

This is how this should be after all. But if this is right here, then what is right for the other link_adjoint?