``` {rate = {covariance = 0.0004716485; weight = 0.0001220703}} Starting timing for: iteration 0 The time was 00:00:07.4534104 for: iteration 0 Training: 0.52367541876136 Testing: 0.278862991322501(9171/9984) ```
Mhhhhhhh...forget it. I just do not feel like going further with this today.

It is impossible to do real programming when one is restless and agitated. Programming requires an investment of time and a tranquil mind.

Without a doubt, I will get it all done tomorrow. It might be the case that the poker test did not work for the same reason `learning1` and `learning2` are not working now. I really need a breather at this point.