``` {rate = {covariance = 0.0004716485; weight = 0.0001220703}} Starting timing for: iteration 0 The time was 00:00:07.3269949 for: iteration 0 Training: 0.523672869779233 Testing: 0.278862844675015(9171/9984) Starting timing for: iteration 1 The time was 00:00:07.2599356 for: iteration 1 Training: 0.269581770882584 Testing: 0.246318095292036(9270/9984) Starting timing for: iteration 2 ```
Interesting. I get the same performance even with the version of the library that I had at the start of the day. The only thing I changed compared to yesterday is plugging in weight sampling.

```
{rate = {covariance = 0.0004716485; weight = 0.0001220703}}
Starting timing for: iteration 0
The time was 00:00:06.7985830 for: iteration 0
Training: 0.520049025479736
Testing: 0.278147141186473(9183/9984)
```

Without weight sampling I get this. Could this be because of some change in the Cuda project?