``` {rate = {covariance = 0.0004716485; weight = 0.0001220703}} Starting timing for: iteration 0 The time was 00:00:07.4876351 for: iteration 0 Training: 0.52367621457252 Testing: 0.278862312149543(9171/9984) ```
It seems not. Is this because I am using tanh instead of ln_relu?

```
{rate = {covariance = 0.0004716485; weight = 0.0001220703}}
Starting timing for: iteration 0
The time was 00:00:07.3440265 for: iteration 0
Training: 0.266546359187207
Testing: 0.0805959385216762(9723/9984)
Starting timing for: iteration 1
The time was 00:00:07.2911276 for: iteration 1
Training: 0.0563786278521586
Testing: 0.0703738412942188(9757/9984)
Starting timing for: iteration 2
```

That seems to be it. `ln_relu` is far more roboust to a variety of learning rates. This is actually a good lesson for me. I'll use `ln_relu` as my main activation. I thought that tanh and ln_relu were even with KFAC, but I was wrong.

```
{rate = {covariance = 0.0004716485; weight = 0.0001220703}}
Starting timing for: iteration 0
The time was 00:00:07.4781524 for: iteration 0
Training: 0.79713591070384
Testing: 0.306425717539894(9145/9984)
Starting timing for: iteration 1
The time was 00:00:07.3315810 for: iteration 1
Training: 0.390488977360929
Testing: 0.259556305451462(9228/9984)
Starting timing for: iteration 2
The time was 00:00:07.3024585 for: iteration 2
Training: 0.358788572140356
Testing: 0.244664292997466(9253/9984)
Starting timing for: iteration 3
```

I expected that it would work in master as soon as I switched back, but this is what I get. Hmmm...let me repeat the experiment from earlier now that I've fixed the activation.