Let me finally start.
Let me backtrack to an earlier commit so I can see what kind of performance does tanh with two layers and -9 LR get me.

```
{rate = {covariance = 0.004978752; weight = 0.001953125}}
Starting timing for: iteration 0
The time was 00:00:07.5507980 for: iteration 0
Training: 0.370661178393512
Testing: 0.237976058174331(9461/9984)
Starting timing for: iteration 1
The time was 00:00:07.5088353 for: iteration 1
Training: 0.183481400468052
Testing: 0.253260000075128(9550/9984)
Starting timing for: iteration 2
The time was 00:00:07.5193915 for: iteration 2
Training: 0.0996288961115271
Testing: 0.208198111992471(9647/9984)
Starting timing for: iteration 3
```

Seems decent enough as a baselline. It really makes me appreciate the `ln_relu` activation.

```
{rate = {covariance = 0.004978752; weight = 0.001953125}}
Starting timing for: iteration 0
The time was 00:00:07.5431575 for: iteration 0
Training: 0.370604540962048
Testing: 0.237262629435207(9463/9984)
Starting timing for: iteration 1
The time was 00:00:07.5449145 for: iteration 1
Training: 0.182734143833926
Testing: 0.207025360256338(9618/9984)
Starting timing for: iteration 2
The time was 00:00:07.6702734 for: iteration 2
Training: 0.0980000675539735
Testing: 0.182288813085623(9686/9984)
Starting timing for: iteration 3
```

Ohhh...it seems the new library does work. There is just something wrong with `ln_relu`. I am really lucky to be able to uncover this so quickly. Let me switch to `ln_relu` for comparison.

```
{rate = {covariance = 0.004978752; weight = 0.001953125}}
Starting timing for: iteration 0
The time was 00:00:07.5464407 for: iteration 0
Training: 0.545498326166063
Testing: 0.31824356641095(9099/9984)
Starting timing for: iteration 1
The time was 00:00:07.5243058 for: iteration 1
Training: 0.521801739342065
Testing: 0.328039392294625(9121/9984)
Starting timing for: iteration 2
```

Based on how this is going, I get the feeling that there is something wrong with gradient propagation in `init_seq`.