"10:05am. Let me move to testing again. I think this should fix it. Yeah, it is almost a given that I am going to end up in trouble once I start putting explicit if statements. I was really rushing it too much yesterday.
10:10am.

```
{rate = {covariance = 0.004978752; weight = 0.001953125}}
Starting timing for: iteration 0
The time was 00:00:07.4313477 for: iteration 0
Training: 0.292786391893736
Testing: 0.190334516675928(9555/9984)
Starting timing for: iteration 1
The time was 00:00:07.3632640 for: iteration 1
Training: 0.0881429088008928
Testing: 0.144671826321999(9691/9984)
Starting timing for: iteration 2
The time was 00:00:07.3690944 for: iteration 2
Training: 0.0322848362601355
Testing: 0.126705171632355(9728/9984)
Starting timing for: iteration 3
The time was 00:00:07.3787847 for: iteration 3
Training: 0.0115624739118696
Testing: 0.113539422500643(9742/9984)
Starting timing for: iteration 4
The time was 00:00:07.3784359 for: iteration 4
Training: 0.00574121541999414
Testing: 0.107907221538881(9776/9984)
Starting timing for: iteration 5
The time was 00:00:07.3754436 for: iteration 5
Training: 0.00197647277671297
Testing: 0.101292729493885(9779/9984)
```

Yeah, this is it! I figured out the bug and fixed it. Let me try -13 learning rate and then I'll move to the char RNN test.

10:20am.

```
{rate = {covariance = 0.0004716485; weight = 0.0001220703}}
Starting timing for: iteration 0
The time was 00:00:07.3442783 for: iteration 0
Training: 0.266546359187207
Testing: 0.0805959385216762(9723/9984)
Starting timing for: iteration 1
The time was 00:00:07.2626512 for: iteration 1
Training: 0.0563779221758103
Testing: 0.07037087686536(9757/9984)
Starting timing for: iteration 2
The time was 00:00:07.2473172 for: iteration 2
Training: 0.0245984095926337
Testing: 0.066889888181975(9778/9984)
```

Yeah, it works great now. Let me try the char RNN."