``` The learning_rate is 2 ** -9 ------ The CudaRandom pseudorandom seed is 0 Starting timing for: iteration 0 Winrate is 210 and 790 out of 1000. The time was 00:00:01.4108282 for: iteration 0 Starting timing for: iteration 1 Winrate is 222 and 778 out of 1000. The time was 00:00:01.3421252 for: iteration 1 Starting timing for: iteration 2 Winrate is 240 and 760 out of 1000. The time was 00:00:01.3609526 for: iteration 2 Starting timing for: iteration 3 Winrate is 256 and 744 out of 1000. The time was 00:00:01.3393156 for: iteration 3 Starting timing for: iteration 4 Winrate is 214 and 786 out of 1000. The time was 00:00:01.3998572 for: iteration 4 Starting timing for: iteration 5 ```
Let me try a lower LR. I think I will get the same result.

```
The learning_rate is 2 ** -11
------
The CudaRandom pseudorandom seed is 0
Starting timing for: iteration 0
Winrate is 210 and 790 out of 1000.
The time was 00:00:01.4387553 for: iteration 0
Starting timing for: iteration 1
Winrate is 222 and 778 out of 1000.
The time was 00:00:01.3309280 for: iteration 1
Starting timing for: iteration 2
Winrate is 240 and 760 out of 1000.
The time was 00:00:01.3570479 for: iteration 2
Starting timing for: iteration 3
Winrate is 256 and 744 out of 1000.
The time was 00:00:01.3444868 for: iteration 3
Starting timing for: iteration 4
Winrate is 214 and 786 out of 1000.
The time was 00:00:01.3825758 for: iteration 4
```

Yeah, getting exactly the same result despite the different learning rate is 100% proof that the net is not being optimized.