"10:50am. Yeah, it works now. Great!
Let me try -9 and then I'll try the RNN with weight sampling.

```
The learning_rate is 2 ** -9
------
The CudaRandom pseudorandom seed is 0
Starting timing for: iteration 0
Winrate is 311 and 689 out of 1000.
The time was 00:00:02.4195743 for: iteration 0
Starting timing for: iteration 1
Winrate is 519 and 481 out of 1000.
The time was 00:00:03.1002213 for: iteration 1
Starting timing for: iteration 2
Winrate is 624 and 376 out of 1000.
The time was 00:00:03.0843369 for: iteration 2
Starting timing for: iteration 3
Winrate is 669 and 331 out of 1000.
The time was 00:00:02.7428088 for: iteration 3
Starting timing for: iteration 4
Winrate is 638 and 362 out of 1000.
The time was 00:00:02.7652113 for: iteration 4
Starting timing for: iteration 5
Winrate is 655 and 345 out of 1000.
The time was 00:00:02.5726331 for: iteration 5
Starting timing for: iteration 6
Winrate is 653 and 347 out of 1000.
The time was 00:00:02.6252148 for: iteration 6
Starting timing for: iteration 7
Winrate is 671 and 329 out of 1000.
The time was 00:00:02.5512138 for: iteration 7
Starting timing for: iteration 8
Winrate is 647 and 353 out of 1000.
The time was 00:00:02.3780246 for: iteration 8
Starting timing for: iteration 9
Winrate is 654 and 346 out of 1000.
The time was 00:00:02.6353142 for: iteration 9
Starting timing for: iteration 10
Winrate is 663 and 337 out of 1000.
The time was 00:00:02.4577553 for: iteration 10
Starting timing for: iteration 11
Winrate is 667 and 333 out of 1000.
The time was 00:00:02.4210591 for: iteration 11
Starting timing for: iteration 12
Winrate is 662 and 338 out of 1000.
The time was 00:00:02.4259143 for: iteration 12
Starting timing for: iteration 13
Winrate is 644 and 356 out of 1000.
The time was 00:00:02.4045664 for: iteration 13
Starting timing for: iteration 14
Winrate is 654 and 346 out of 1000.
The time was 00:00:02.3885028 for: iteration 14
```

Hmmm, the linear net with -9 did not reach over 70%. Agh. Well, forget that, let me try the RNN.

```
The learning_rate is 2 ** -11
------
The CudaRandom pseudorandom seed is 0
Starting timing for: iteration 0
Winrate is 357 and 643 out of 1000.
The time was 00:00:06.0718542 for: iteration 0
Starting timing for: iteration 1
Winrate is 657 and 343 out of 1000.
The time was 00:00:07.5650782 for: iteration 1
Starting timing for: iteration 2
Winrate is 694 and 306 out of 1000.
The time was 00:00:07.3543865 for: iteration 2
Starting timing for: iteration 3
Winrate is 698 and 302 out of 1000.
The time was 00:00:07.6501565 for: iteration 3
Starting timing for: iteration 4
Winrate is 721 and 279 out of 1000.
The time was 00:00:07.3417334 for: iteration 4
Starting timing for: iteration 5
Winrate is 727 and 273 out of 1000.
The time was 00:00:06.8765025 for: iteration 5
Starting timing for: iteration 6
Winrate is 720 and 280 out of 1000.
The time was 00:00:07.3143558 for: iteration 6
Starting timing for: iteration 7
Winrate is 732 and 268 out of 1000.
The time was 00:00:07.0785602 for: iteration 7
Starting timing for: iteration 8
Winrate is 709 and 291 out of 1000.
The time was 00:00:07.0982402 for: iteration 8
Starting timing for: iteration 9
Winrate is 746 and 254 out of 1000.
The time was 00:00:07.5843251 for: iteration 9
Starting timing for: iteration 10
Winrate is 739 and 261 out of 1000.
The time was 00:00:07.0908648 for: iteration 10
Starting timing for: iteration 11
Winrate is 735 and 265 out of 1000.
The time was 00:00:06.9872350 for: iteration 11
Starting timing for: iteration 12
Winrate is 728 and 272 out of 1000.
The time was 00:00:07.3887922 for: iteration 12
Starting timing for: iteration 13
Winrate is 711 and 289 out of 1000.
The time was 00:00:07.0723415 for: iteration 13
Starting timing for: iteration 14
Winrate is 758 and 242 out of 1000.
The time was 00:00:07.3087199 for: iteration 14
```

Yeah, this is pretty good. It does seem like the RNN is making a noticeable difference. Let me increase the LR. I want to see if weight sampling makes it more roboust at the core.

10:55am. I'll also add weight sampling to the AC layer.

11:25am. Dad made breakfast so I had it early with him today. Let me run the test again.

```
The learning_rate is 2 ** -10
------
The CudaRandom pseudorandom seed is 0
Starting timing for: iteration 0
Winrate is 507 and 493 out of 1000.
The time was 00:00:06.2819968 for: iteration 0
Starting timing for: iteration 1
Winrate is 671 and 329 out of 1000.
The time was 00:00:06.7685022 for: iteration 1
Starting timing for: iteration 2
Winrate is 673 and 327 out of 1000.
The time was 00:00:06.4625807 for: iteration 2
Starting timing for: iteration 3
Winrate is 699 and 301 out of 1000.
The time was 00:00:06.9122230 for: iteration 3
Starting timing for: iteration 4
Winrate is 698 and 302 out of 1000.
The time was 00:00:06.3343135 for: iteration 4
Starting timing for: iteration 5
Winrate is 688 and 312 out of 1000.
The time was 00:00:06.2954665 for: iteration 5
Starting timing for: iteration 6
Winrate is 678 and 322 out of 1000.
The time was 00:00:06.8711660 for: iteration 6
Starting timing for: iteration 7
Winrate is 727 and 273 out of 1000.
The time was 00:00:07.8281418 for: iteration 7
Starting timing for: iteration 8
Winrate is 727 and 273 out of 1000.
The time was 00:00:07.7003430 for: iteration 8
Starting timing for: iteration 9
Winrate is 715 and 285 out of 1000.
The time was 00:00:08.4160582 for: iteration 9
Starting timing for: iteration 10
Winrate is 744 and 256 out of 1000.
The time was 00:00:07.9414258 for: iteration 10
Starting timing for: iteration 11
Winrate is 738 and 262 out of 1000.
The time was 00:00:07.5060483 for: iteration 11
Starting timing for: iteration 12
Winrate is 735 and 265 out of 1000.
The time was 00:00:07.7871758 for: iteration 12
Starting timing for: iteration 13
Winrate is 713 and 287 out of 1000.
The time was 00:00:07.3618466 for: iteration 13
Starting timing for: iteration 14
Winrate is 753 and 247 out of 1000.
The time was 00:00:07.3348461 for: iteration 14
```

This is good. Let me try adding weight sampling to the AC net...as soon as I finish this fruit.

11:55am. https://mangadex.org/chapter/7713/2

Let me stop reading this manga here. It is time to resume this, though there is not much work left.

The -10 run is especially good, if it was the old net without sampling it would have never recovered from 678 on iteration 6. Let me add weight sampling to the AC."