"12pm. https://arxiv.org/abs/1812.03365 Neuromodulated Learning in Deep Neural Networks
For a bit I thought this would be related to differentiable plasticity, but it is not. I'll take a look at it later.

12:05pm.

```
            apply = inl {d with weights={weights} input} s ->
                inl weights =
                    match d with
                    | {state={weights}} -> weights
                    | _ -> weight_sample weights s

                inl apply =
                    inm data = concat {bias=one; input}
                    inm out = matmult_stream {weights with data}
                    succ (View.split out)

                inl {out bck} = apply s
                {out bck state={weights}}
```

This should do the trick. Let me give it a shot.

```
The learning_rate is 2 ** -10
------
The CudaRandom pseudorandom seed is 0
Starting timing for: iteration 0
Winrate is 482 and 518 out of 1000.
The time was 00:00:07.8062694 for: iteration 0
Starting timing for: iteration 1
Winrate is 621 and 379 out of 1000.
The time was 00:00:07.7109927 for: iteration 1
Starting timing for: iteration 2
Winrate is 678 and 322 out of 1000.
The time was 00:00:07.9051805 for: iteration 2
Starting timing for: iteration 3
Winrate is 708 and 292 out of 1000.
The time was 00:00:07.9590331 for: iteration 3
Starting timing for: iteration 4
Winrate is 712 and 288 out of 1000.
The time was 00:00:08.1882833 for: iteration 4
Starting timing for: iteration 5
Winrate is 728 and 272 out of 1000.
The time was 00:00:07.9923311 for: iteration 5
Starting timing for: iteration 6
Winrate is 729 and 271 out of 1000.
The time was 00:00:07.7855949 for: iteration 6
Starting timing for: iteration 7
Winrate is 732 and 268 out of 1000.
The time was 00:00:07.7540026 for: iteration 7
Starting timing for: iteration 8
Winrate is 700 and 300 out of 1000.
The time was 00:00:07.5384142 for: iteration 8
Starting timing for: iteration 9
Winrate is 713 and 287 out of 1000.
The time was 00:00:07.7254691 for: iteration 9
Starting timing for: iteration 10
Winrate is 728 and 272 out of 1000.
The time was 00:00:07.3905983 for: iteration 10
Starting timing for: iteration 11
Winrate is 729 and 271 out of 1000.
The time was 00:00:07.2391860 for: iteration 11
Starting timing for: iteration 12
Winrate is 690 and 310 out of 1000.
The time was 00:00:07.0330966 for: iteration 12
Starting timing for: iteration 13
Winrate is 724 and 276 out of 1000.
The time was 00:00:07.4962040 for: iteration 13
Starting timing for: iteration 14
Winrate is 714 and 286 out of 1000.
The time was 00:00:07.3610925 for: iteration 14
```

Hmmmm, it seems the noise is preventing it from trully converging. As expected. Let me try -11 just to make sure and then I will remove weight sampling from the final layer."