``` The learning_rate is 2 ** -10 ------ The CudaRandom pseudorandom seed is 0 Starting timing for: iteration 0 Winrate is 306 and 694 out of 1000. The time was 00:00:11.1922939 for: iteration 0 Starting timing for: iteration 1 Winrate is 365 and 635 out of 1000. The time was 00:00:11.4737358 for: iteration 1 Starting timing for: iteration 2 Winrate is 381 and 619 out of 1000. The time was 00:00:11.2613301 for: iteration 2 Starting timing for: iteration 3 Winrate is 359 and 641 out of 1000. The time was 00:00:11.3125830 for: iteration 3 ```
Hmm, maybe I need to lower it to -12 instead.

```
The learning_rate is 2 ** -12
------
The CudaRandom pseudorandom seed is 0
Starting timing for: iteration 0
Winrate is 314 and 686 out of 1000.
The time was 00:00:12.1561277 for: iteration 0
Starting timing for: iteration 1
Winrate is 618 and 382 out of 1000.
The time was 00:00:15.8708629 for: iteration 1
Starting timing for: iteration 2
Winrate is 694 and 306 out of 1000.
The time was 00:00:16.6617983 for: iteration 2
Starting timing for: iteration 3
Winrate is 743 and 257 out of 1000.
The time was 00:00:16.6392906 for: iteration 3
Starting timing for: iteration 4
Winrate is 719 and 281 out of 1000.
The time was 00:00:16.5429154 for: iteration 4
Starting timing for: iteration 5
Winrate is 748 and 252 out of 1000.
The time was 00:00:16.2520927 for: iteration 5
Starting timing for: iteration 6
Winrate is 735 and 265 out of 1000.
The time was 00:00:15.9657758 for: iteration 6
Starting timing for: iteration 7
Winrate is 734 and 266 out of 1000.
The time was 00:00:16.0872701 for: iteration 7
Starting timing for: iteration 8
Winrate is 728 and 272 out of 1000.
The time was 00:00:16.7035904 for: iteration 8
Starting timing for: iteration 9
Winrate is 741 and 259 out of 1000.
The time was 00:00:17.4636970 for: iteration 9
Starting timing for: iteration 10
Winrate is 744 and 256 out of 1000.
The time was 00:00:15.7039466 for: iteration 10
Starting timing for: iteration 11
Winrate is 726 and 274 out of 1000.
The time was 00:00:15.9381280 for: iteration 11
Starting timing for: iteration 12
Winrate is 747 and 253 out of 1000.
The time was 00:00:16.2623575 for: iteration 12
Starting timing for: iteration 13
Winrate is 735 and 265 out of 1000.
The time was 00:00:16.5453740 for: iteration 13
Starting timing for: iteration 14
Winrate is 741 and 259 out of 1000.
The time was 00:00:16.0684732 for: iteration 14
```

Yeah, that was the winning ticket. I have no doubt this is optimal. That having said, unlike with Mnist the range allowable LRs are really tight here. I should implement hypergradient descent.

12:35pm. Do I feel like doing that right now?

Yeah, it is not like I do not have the time. I really should just put all the good ideas in. I have no reason not to.

Let me do it. I want to leave the library in a pristine condition before I take a break from this for a bit.

...Actually depending on how well hypergradient descent works, I might in fact go for it with this player. I have no reason not to try Reptile again either.