"Ok, now KFAC's matrix inverse is not blowing up. But this thing is quite unstable. It is not what I would expect given the paper.
Let me turn off KFAC for a bit. I'll study how it behaves with standard SGD.

```
{rate = {covariance = 0.85; weight = 0.0001220703}}
Starting timing for: iteration 0
3.1E-05
3.1E-05
3.1E-05
3.1E-05
3.1E-05
3.1E-05
-0.03726847
-0.002661619
0.02072562
-0.03530096
4.292709
-2.155775
The time was 00:00:00.0639208 for: iteration 0
Training: NaN
Training diverged. Aborting...
```

It is deeply unstable. There is something wrong with all of this.

3:30pm. I am thinking. Weight sampling might be making it unstable. Having layerwise learning rates might be a problem as well.

Hence, I will turn off weight sampling and go to a single linear layer to check what is going on. According to the paper, hypergradient descent should be super solid and here it is acting like a bomb. There is something wrong with all of this."