"3:30pm. I am thinking. Weight sampling might be making it unstable. Having layerwise learning rates might be a problem as well.
Hence, I will turn off weight sampling and go to a single linear layer to check what is going on. According to the paper, hypergradient descent should be super solid and here it is acting like a bomb. There is something wrong with all of this.

```
{rate = {covariance = 0.85; weight = 0.0001220703}}
Starting timing for: iteration 0
3.1E-05
3.1E-05
2.513411
The time was 00:00:00.0194457 for: iteration 0
Training: NaN
Training diverged. Aborting...
```

As a sanity check, let me turn off the LR updating. What will softmax regression get me?

```
Starting timing for: iteration 0
The time was 00:00:00.1947897 for: iteration 0
Training: 1.64992029582843
Testing: 1.21101546440369(8156/9984)
Starting timing for: iteration 1
The time was 00:00:00.1757994 for: iteration 1
Training: 1.04503345349406
Testing: 0.884692384646489(8393/9984)
Starting timing for: iteration 2
The time was 00:00:00.1688467 for: iteration 2
Training: 0.828977809184127
Testing: 0.739219383169443(8527/9984)
Starting timing for: iteration 3
The time was 00:00:00.1688571 for: iteration 3
Training: 0.719533155807573
Testing: 0.656258950630824(8616/9984)
Starting timing for: iteration 4
The time was 00:00:00.1687251 for: iteration 4
Training: 0.652352241560435
Testing: 0.601983570326597(8671/9984)
```

I could better results with a higher learning rate, but this will do.

Could I possibly be updating the learning rates in the wrong direction. I do not think that could be the case...

```
{rate = {covariance = 0.85; weight = 0.0001220703}}
Starting timing for: iteration 0
3.1E-05
3.1E-05
-2.513349
The time was 00:00:00.0195111 for: iteration 0
Training: NaN
Training diverged. Aborting...
```

Yeah, no. I guess I could lower beta even more, but this thing...

Let me print out the adjoints.

3:45pm. I have no idea. Let me try pushing the beta down to -20. That might stabilize it.

After that I'll take a look at the repo for hypergrad descent."