"3:55pm. As a sanity check let me print out just the dot product at every iteration. It is highly unlikely, but there might be something wrong with the redo function.
```
{dot = 7254.398}
3.1E-05
{dot = 5389.741}
3.1E-05
{dot = 7965.951}
3.1E-05
{dot = 8198.847}
3.1E-05
{dot = 8049.7}
3.1E-05
{dot = 6509.378}
3.1E-05
{dot = 7416.023}
3.1E-05
{dot = 5110.632}
3.1E-05
{dot = 6964.609}
3.1E-05
{dot = 8339.661}
```

These dot products are fairly large. 784*10. Essentially this result means that the size of the dot product of the gradients is proportional to the number of weights.

They are this large...really?

Let me reason it out - the inputs are in the [0,1] range. The output gradients are in the [0,1] range. The weight gradients should be in the [0,1] range. Therefore there is very little chance that this dot is correct.

Let me print out the matrices individually."