"3:55pm. As a sanity check let me print out just the dot product at every iteration. It is highly unlikely, but there might be something wrong with the redo function.
```
{dot = 7254.398}
3.1E-05
{dot = 5389.741}
3.1E-05
{dot = 7965.951}
3.1E-05
{dot = 8198.847}
3.1E-05
{dot = 8049.7}
3.1E-05
{dot = 6509.378}
3.1E-05
{dot = 7416.023}
3.1E-05
{dot = 5110.632}
3.1E-05
{dot = 6964.609}
3.1E-05
{dot = 8339.661}
```

These dot products are fairly large. 784*10. Essentially this result means that the size of the dot product of the gradients is proportional to the number of weights.

They are this large...really?

Let me reason it out - the inputs are in the [0,1] range. The output gradients are in the [0,1] range. The weight gradients should be in the [0,1] range. Therefore there is very little chance that this dot is correct.

Let me print out the matrices individually.

4:10pm.

```
[|-1.102111; -4.74727; -1.904146; 3.420217; -1.850648; -0.1283404; -0.8834922; 1.840174; 3.158421; 2.197196|]
    [|-0.3281823; -2.712949; -2.846292; 3.926507; -4.915654; 0.5755731; -1.136655; 0.8155787; 3.395386; 3.226688|]
    [|-0.4641922; 0.2201832; -3.188325; 3.851062; -3.541964; 1.209596; -1.026534; -2.027488; 2.353615; 2.614047|]
```

I forgot something really important. Namely that I am using minibatching. This will be annoying, but let me try setting the minibatch size to 1. I really should have done it on a separate branch after all. I really wish the paper mentioned this.

```
For MNIST one epoch is one full pass through the entire training set of 60,000
images (468.75 iterations with a minibatch size of 128) and for CIFAR-10 one epoch is one full pass
through the entire training set of 50,000 images (390.625 iterations with a minibatch size of 128).
```

Wait, it is using minibatching in the paper.

Ok, what the hell?

No, I can't make any headway like this. Let me try cloning the repo and running the tests that are there.

I honestly did not thing this would be so much work, but it is clear to me that I am missing something important."