``` {rate = {covariance = 0.85; minibatch_size = 128; weight = 0.0001220703}} Starting timing for: iteration 0 The time was 00:00:02.2893059 for: iteration 0 Training: 1.38279721771295 Testing: 0.731331092806963(8747/9984) Starting timing for: iteration 1 The time was 00:00:02.2557007 for: iteration 1 Training: 0.602242902653594 Testing: 0.487247915031054(8925/9984) Starting timing for: iteration 2 The time was 00:00:02.2443245 for: iteration 2 Training: 0.465540958043092 Testing: 0.413502062647006(8998/9984) Starting timing for: iteration 3 The time was 00:00:02.2449357 for: iteration 3 Training: 0.411191044518581 Testing: 0.377270053212459(9041/9984) Starting timing for: iteration 4 The time was 00:00:02.2419572 for: iteration 4 Training: 0.380693216020098 Testing: 0.355383740021632(9086/9984) ```
I think the multiplicative update is a bit slow. Let me increse the LR for it.

```
inl default_learning_rate = epsilon -15
```

Let me also finally set this to -20 like I had intended to for a while now.

```
{rate = {covariance = 0.85; minibatch_size = 128; weight = 0.0078125}}
Starting timing for: iteration 0
The time was 00:00:02.2963530 for: iteration 0
Training: 2.23183414671156
Testing: 2.0859846274058(8189/9984)
Starting timing for: iteration 1
The time was 00:00:02.2578802 for: iteration 1
Training: 1.68541620646277
Testing: 1.18238731760245(8553/9984)
Starting timing for: iteration 2
The time was 00:00:02.2420911 for: iteration 2
Training: 0.881710424127742
Testing: 0.642023093043229(8814/9984)
Starting timing for: iteration 3
The time was 00:00:02.2473152 for: iteration 3
Training: 0.566413456590002
Testing: 0.473487835472975(8937/9984)
Starting timing for: iteration 4
The time was 00:00:02.2548136 for: iteration 4
Training: 0.453581458642187
Testing: 0.402154102730445(9010/9984)
```

It takes it a while to update the LRs. Let me print them out.

```
3.848552E-05
3.849202E-05
3.85292E-05
3.854038E-05
3.854182E-05
3.857057E-05
3.858944E-05
3.86269E-05
3.86542E-05
3.866032E-05
3.866496E-05
3.870503E-05
3.871845E-05
3.871803E-05
The time was 00:00:02.3797729 for: iteration 4
Training: 0.453581458642187
Testing: 0.402154102730445(9010/9984)
```

It really takes it a long time to adjust the learning rate from `2 ** -20` with a high beta.

Let me try the additive again.

```
0.004736668
0.004688381
0.004674191
The time was 00:00:02.4593995 for: iteration 4
Training: 0.339206328368786
Testing: 0.37502695648716(9033/9984)
```

The additive one adjusts it in no time at all. I am really leaning towards using the additive update for everything here. Let me see how it acts in just the first iteration.